requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openvino>=2024.0.0
optimum-intel[openvino,npu]>=1.13.0
transformers>=4.37.0
huggingface-hub>=0.19.0
tokenizers>=0.15.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart==0.0.6
jinja2>=3.1.0
aiofiles>=23.2.0
python-json-logger>=2.0.7
psutil>=5.9.0